requirements.txt
pyyaml
pymodbus